Rename to storage_cleanup
